+++ Sprint 6 Challenge/Sprint 6 Challenge/Resources/AppDelegate.swift
@@ -1,0 +1,44 @@
+//
+//  AppDelegate.swift
+//  Sprint 6 Challenge
+//
+//  Created by Spencer Curtis on 8/29/18.
+//  Copyright © 2018 Lambda School. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        Appearance.setAppearance()
+        return true
+    }
+
+    func applicationWillResignActive(_ application: UIApplication) {
+    
+    }
+
+    func applicationDidEnterBackground(_ application: UIApplication) {
+ 
+    }
+
+    func applicationWillEnterForeground(_ application: UIApplication) {
+
+    }
+
+    func applicationDidBecomeActive(_ application: UIApplication) {
+
+    }
+
+    func applicationWillTerminate(_ application: UIApplication) {
+
+    }
+
+
+}
+
