+++ Sprint 6 Challenge/Sprint 6 Challenge/View Controller/ViewController.swift
@@ -30,7 +30,7 @@ class ViewController: UIViewController {
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
+        
     }
